Refactors Constants, creating additional contant for rsa machine key paths
