+++ src/Encrypt.Config.ConsoleHost/Constants/WellKnownCommandArguments.cs
