+++ src/Encrypt.Config.ConsoleHost/Constants/WellKnownCommands.cs
@@ -1,6 +1,7 @@
 namespace Encrypt.Config.ConsoleHost {
     public class WellKnownCommands
     {
+        public const string CREATE_KEYS = "create";
         public const string EXPORT_KEY = "export";
         public const string ENCRYPT_FILE = "encrypt";
         public const string DECRYPT_FILE = "decrypt";
