+++ src/Encrypt.Config.Encryption/Constants/WellKnownPaths.cs
@@ -1,0 +1,7 @@
+namespace Encrypt.Config.Encryption.Constants {
+    public class WellKnownPaths
+    {
+        public const string RSA_MACHINEKEYS = @"C:\ProgramData\Microsoft\Crypto\RSA\MachineKeys";
+
+    }
+}
